Fix #34 Arrumando duplicidade de SearchHome e ajustando onde e importada e pequenos ajustes
