+++ src/screens/DeputyProfileScreen.js
@@ -51,6 +51,7 @@ function DeputyProfileScreen() {
         <Row className="space" />
         <SpentData />
         <Row className="space" />
+        <News />
       </Container>
       <Row className="space" />
     </main>
